@@ -1,6 +1,6 @@
 cmake_minimum_required(VERSION 2.8.12)
 project(PackageTest CXX)
-set(CMAKE_CXX_STANDARD 14)
+set(CMAKE_CXX_STANDARD 17)
 set(CMAKE_CXX_STANDARD_REQUIRED on)
 
 find_package(wfut CONFIG REQUIRED)
