@@ -24,9 +24,9 @@ class LibWFUTConan(ConanFile):
     package_type = "library"
 
     def requirements(self):
-        self.requires("sigc++/2.10.0@worldforge", transitive_headers=True)
+        self.requires("libsigcpp/3.0.7", transitive_headers=True)
         self.requires("zlib/1.2.13")
-        self.requires("libcurl/7.87.0")
+        self.requires("libcurl/7.88.1")
 
     def export(self):
         git = Git(self, self.recipe_folder)
@@ -38,7 +38,7 @@ class LibWFUTConan(ConanFile):
 
     def export_sources(self):
         folder = os.path.join(self.recipe_folder, "../..")
-        copy(self, "*", folder, self.export_sources_folder, excludes=["*build*"])
+        copy(self, "*", folder, self.export_sources_folder, excludes=["build"])
 
     def config_options(self):
         if self.settings.os == "Windows":
